[10.0][IMP] cagnotte_*: Improve tests, module description, ...
