+++ cagnotte_base/README.rst
@@ -2,21 +2,76 @@
 Cagnotte Base
 =============
 
-This module manages cagnotte :
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-* define product on cagnotte to fill it
-* define journal on cagnotte to empty it
-* on account move lines assign cagnotte
-* When cagnotte is empty, it is not possible to use it
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-acsone%2Facsone--addons-lightgray.png?logo=github
+    :target: https://github.com/acsone/acsone-addons/tree/10.0/cagnotte_base
+    :alt: acsone/acsone-addons
+
+|badge1| |badge2| |badge3| 
+
+This module allows to manage cagnottes (financial reserve).
+
+
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Usage
+=====
+
+* Go to Accounting > Cagnotte
+* Define first a Cagnotte type.
+* For move lines generation, please:
+    * Choose a sequence
+    * Choose an account
+    * Choose a product
+    * Choose a journal
+
+* Go to Accounting > Cagnotte > Cagnotte
+* Create a Cagnotte,
+* When cagnotte is empty, it is not possible to use it anymore
+
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/acsone/acsone-addons/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/acsone/acsone-addons/issues/new?body=module:%20cagnotte_base%0Aversion:%2010.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
 
 Credits
 =======
 
-Maintainer
-----------
+Authors
+~~~~~~~
 
-.. image:: https://www.acsone.eu/logo.png
-   :alt: ACSONE SA/NV
-   :target: http://www.acsone.eu
+* ACSONE SA/NV
 
-This module is maintained by ACSONE SA/NV.
+Contributors
+~~~~~~~~~~~~
+
+* Laetitia Gangloff <laetitia.gangloff@acsone.eu>
+* Cedric Pigeon <cedric.pigeon@acsone.eu>
+* Denis Roussel <denis.roussel@acsone.eu>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `acsone/acsone-addons <https://github.com/acsone/acsone-addons/tree/10.0/cagnotte_base>`_ project on GitHub.
+
+You are welcome to contribute.
