--- cagnotte_base/demo/account_cagnotte_demo.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="cagnotte" model="account.account">
-        <field name="code">X4404</field>
-        <field name="name">Cagnotte - (test)</field>
-        <field name="parent_id" ref="account.cas"/>
-        <field name="type">other</field>
-        <field name="user_type" ref="account.data_account_type_income"/>
-    </record>
-    <record id="sequence_cagnotte_journal" model="ir.sequence">
-        <field name="name">Account Default Cagnotte Journal</field>
-        <field eval="3" name="padding"/>
-        <field name="prefix">CAG/%(year)s/</field>
-    </record>
-    <record id="cagnotte_journal" model="account.journal">
-        <field name="name">Cagnotte Journal - (test)</field>
-        <field name="code">CAG</field>
-        <field name="type">bank</field>
-        <field name="sequence_id" ref="sequence_cagnotte_journal"/>
-        <field name="default_debit_account_id" ref="cagnotte"/>
-        <field name="default_credit_account_id" ref="cagnotte"/>
-    </record>
-    <record id="product_product_cagnotte" model="product.product">
-        <field name="name">Cagnotte</field>
-        <field name="standard_price">15</field>
-        <field name="type">service</field>
-        <field name="property_account_income" ref="cagnotte"/>
-    </record>
-    <record id="sequence_cagnotte_type" model="ir.sequence">
-        <field name="name">Cagnotte type</field>
-        <field eval="3" name="padding"/>
-        <field name="prefix">CAG-</field>
-    </record>
-    <record id="cagnotte_type" model="cagnotte.type">
-        <field name="name">Cagnotte Type</field>
-        <field name="sequence_id" ref="sequence_cagnotte_type"/>
-        <field name="account_id" ref="cagnotte"/>
-        <field name="journal_id" ref="cagnotte_journal"/>
-        <field name="product_id" ref="product_product_cagnotte"/>
-    </record>
-</odoo>
