+++ cagnotte_base/models/__init__.py
@@ -1,3 +1,4 @@
-# -*- coding: utf-8 -*-
-
-from . import cagnotte
+from . import cagnotte_type
+from . import account_cagnotte
+from . import account_move_line
+from . import account_invoice
