+++ cagnotte_base/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Laetitia Gangloff <laetitia.gangloff@acsone.eu>
+* Cedric Pigeon <cedric.pigeon@acsone.eu>
+* Denis Roussel <denis.roussel@acsone.eu>
