+++ cagnotte_base/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+This module allows to manage cagnottes (financial reserve).
+
+
