+++ cagnotte_base/readme/USAGE.rst
@@ -1,0 +1,12 @@
+* Go to Accounting > Cagnotte
+* Define first a Cagnotte type.
+* For move lines generation, please:
+    * Choose a sequence
+    * Choose an account
+    * Choose a product
+    * Choose a journal
+
+* Go to Accounting > Cagnotte > Cagnotte
+* Create a Cagnotte,
+* When cagnotte is empty, it is not possible to use it anymore
+
