+++ cagnotte_base/security/cagnotte_base_security.xml
@@ -1,17 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data noupdate="1">
-        <record id="cagnotte_type_comp_rule" model="ir.rule">
-            <field name="name">Cagnotte Type multi company</field>
-            <field name="model_id" ref="model_cagnotte_type"/>
-            <field name="global" eval="True"/>
-            <field name="domain_force">[('company_id','child_of',[user.company_id.id])]</field>
-        </record>
-        <record id="account_cagnotte_comp_rule" model="ir.rule">
-            <field name="name">Cagnotte multi company</field>
-            <field name="model_id" ref="model_account_cagnotte"/>
-            <field name="global" eval="True"/>
-            <field name="domain_force">[('cagnotte_type_id.company_id','child_of',[user.company_id.id])]</field>
-        </record>
-    </data>
+<odoo noupdate="1">
+    <record id="cagnotte_type_comp_rule" model="ir.rule">
+        <field name="name">Cagnotte Type multi company</field>
+        <field name="model_id" ref="model_cagnotte_type"/>
+        <field name="global" eval="True"/>
+        <field name="domain_force">[('company_id','child_of',[user.company_id.id])]</field>
+    </record>
+    <record id="account_cagnotte_comp_rule" model="ir.rule">
+        <field name="name">Cagnotte multi company</field>
+        <field name="model_id" ref="model_account_cagnotte"/>
+        <field name="global" eval="True"/>
+        <field name="domain_force">[('cagnotte_type_id.company_id','child_of',[user.company_id.id])]</field>
+    </record>
+
 </odoo>
