+++ cagnotte_base/tests/__init__.py
@@ -1,3 +1,2 @@
-# -*- coding: utf-8 -*-
-
+from . import common
 from . import test_cagnotte
