+++ cagnotte_coupon/tests/test_cagnotte.py
@@ -1,20 +1,10 @@
 # -*- coding: utf-8 -*-
 # © 2015  Laetitia Gangloff, Acsone SA/NV (http://www.acsone.eu)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-import odoo.tests.common as common
-from odoo.addons.cagnotte_base.tests.test_cagnotte import load_file
+from odoo.addons.cagnotte_base.tests.common import CagnotteCommon
 
 
-class TestCagnotte(common.TransactionCase):
-
-    def setUp(self):
-        super(TestCagnotte, self).setUp()
-        load_file(
-            self.cr,
-            'cagnotte_base',
-            'tests/data/',
-            'account_cagnotte_data.xml')
+class TestCagnotte(CagnotteCommon):
 
     def test_cagnotte(self):
         """ Create cagnotte
