+++ cagnotte_partner/README.rst
