+++ cagnotte_partner/tests/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
-
+from . import common
 from . import test_cagnotte
